H.Con.Res. 414 Introduced in House
Expressing the sense of the Congress that, as Congress recognizes the 50th anniversary of the Brown v. Board of Education decision, all Americans are encouraged to observe this anniversary with a commitment to continuing and building on the legacy of Brown.

Version:  Introduced in House (ih)
Date:     2004-05-04
Congress: 108

Source: https://www.govinfo.gov/content/pkg/BILLS-108hconres414ih/xml/BILLS-108hconres414ih.xml

Sponsored-By: C000714
Cosponsor-Count: 63